Moved verification to its own class
